+++ ext/avif/CHANGELOG.md
@@ -6,6 +6,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.4.1] - 2019-10-17
+### Added
+- Added `containerDepth` to avifDecoder for surfacing 10bpc/12bpc flags from av1C boxes, if present
+- Added `avifCodecVersions()` for getting version strings of internal AV1 codecs
+
+### Changed
+- Fixed warning with CHECK macro (additional semicolon)
+
 ## [0.4.0] - 2019-10-02
 ### Added
 - exposed util functions: `avifFullToLimitedY`, `avifFullToLimitedUV`, `avifLimitedToFullY`, `avifLimitedToFullUV`, `avifPrepareReformatState`
@@ -185,7 +193,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Constants `AVIF_VERSION`, `AVIF_VERSION_MAJOR`, `AVIF_VERSION_MINOR`, `AVIF_VERSION_PATCH`
 - `avifVersion()` function
 
-[Unreleased]: https://github.com/AOMediaCodec/libavif/compare/v0.4.0...HEAD
+[Unreleased]: https://github.com/AOMediaCodec/libavif/compare/v0.4.1...HEAD
+[0.4.1]: https://github.com/AOMediaCodec/libavif/compare/v0.4.0...v0.4.1
 [0.4.0]: https://github.com/AOMediaCodec/libavif/compare/v0.3.11...v0.4.0
 [0.3.11]: https://github.com/AOMediaCodec/libavif/compare/v0.3.10...v0.3.11
 [0.3.10]: https://github.com/AOMediaCodec/libavif/compare/v0.3.9...v0.3.10